idpf: refactor Tx completion routines
This patch adds a mechanism to guard against stashing partial packets
into the hash table. This makes the driver more robust, leads to more
efficient decision making when cleaning.

Doon't stash partial packets. This can happen when an RE completion is
received in flow scheduling mode, or when an out of order RS completion
is received. The first buffer with the skb is stashed, but some or all
of its frags are not because the stack is out of reserve buffers. This
leaves the ring in a weird state since the frags are still on the ring.

Use the field to track the number of fragments/ tx_bufs representing the
packet. The clean routines check to make sure there are enough reserve
buffers on the stack before stashing any part of the packet. If there
are not, next_to_clean is left pointing to the first buffer of the
packet that failed to be stashed. This leaves the whole packet on the
ring, and the next time around, cleaning will start from this packet.

An RS completion is still expected for this packet in either case. So
instead of being cleaned from the hash table, it will be cleaned from
the ring directly.  This should all still be fine since the DESC_UNUSED
and BUFS_UNUSED will reflect the state of the ring. If we ever fall
below the thresholds, the TXQ will still be stopped, giving the
completion queue time to catch up.  This may lead to stopping the queue
more frequently, but it guarantees the TX ring will always be in a good
state.

Also, always use the idpf_tx_splitq_clean function to clean descriptors,
i.e. use it from clean_buf_ring as well. This way we avoid duplicating
the logic and make sure we're using the same reserve buffers guard rail.

This does require a switch from the s16 next_to_clean overflow
descriptor ring wrap calculation to u16 and the normal ring size check.

Signed-off-by: Joshua Hay <joshua.a.hay@intel.com>
Signed-off-by: Alexander Lobakin <aleksander.lobakin@intel.com>