idpf-linux: fix memory leaks in xdp and soft_reset
The XDP initialization and soft_reset can allocate queues before
opening the vport to be able to request and configure interrupts
for that queues.
When the link is up during such an operation, 'idpf_vport_open()'
is called with a parameter to skip the redundant resources allocation
while the vport is being opened.
However, when the link is down, a redundant allocation of queues can
take place after user explicitly will set the link up, causing
the memory leak.

Fix that "link down" scenario by freeing the redundant allocations
in XDP initialization and in soft_reset when the link is down, so
the vport should remain closed (with no queues allocated).

Example reproduction scenario for XDP:
1. Set the link down.
2. Load the XDP program.
3. Unload the driver (rmmod idpf).

Reproduction scenarion for soft_reset:
1. Set the link down.
2. Change the queue number (using ethtool).
3. Set the link up again.

In both scenarios, (when the fix is not applied), kernel memory leak
detector should show the memory leak in queue allocation.

Signed-off-by: Michal Kubiak <michal.kubiak@intel.com>