ice: disable flow director [for POC only]
Since the ice driver uses the same Rx path for handling regular NAPI Rx
queues as well as control VSI flow director queues, the page_pool could
not be applied in a generic way.
Skip that design issue temporarily, just to avoid the POC crashes, by
disabling the flow director completely.

The issue has to be addressed after we decide to upstream that page_pool
solution.

Signed-off-by: Michal Kubiak <michal.kubiak@intel.com>