ice: fix rebuilding the Tx scheduler tree for large queue counts
The current implementation of the Tx scheduler allows the tree to be
rebuilt as the user adds more Tx queues to the VSI. In such a case,
additional child nodes are added to the tree to support the new number
of queues.
Unfortunately, this algorithm does not take into account that the limit
of the VSI support node may be exceeded, so an additional node in the
VSI layer may be required to handle all the requested queues.

Such a scenario occurs when adding XDP Tx queues on machines with many
CPUs (e.g. over ~300). Although the driver still respects the queue
limit returned by the FW, the Tx scheduler was unable to add those
queues to its tree and returned the following error:

    Failed VSI LAN queue config for XDP, error: -5

Fix this problem by extending the tree rebuild algorithm to check if the
current VSI node can support the requested number of queues. If it
cannot, create as many additional VSI support nodes as necessary to
handle all the required Tx queues.
Also, modify the node count calculation algorithm to skip the queues
that already exist in the tree. By the way, clarify the comment
describing the tree update algorithm.

Signed-off-by: Michal Kubiak <michal.kubiak@intel.com>